+++ example/lib/injection.dart
@@ -1,0 +1,15 @@
+import 'package:dart_cqrs/dart_cqrs.dart';
+import 'package:get_it/get_it.dart';
+import 'package:injectable/injectable.dart';
+
+import 'injection.config.dart';
+
+final getIt = GetIt.instance;
+
+@InjectableInit(
+  allowMultipleRegistrations: true,
+  externalPackageModulesBefore: [
+    ExternalModule(DartCqrsPackageModule),
+  ],
+)
+Future<void> configureDependencies() => getIt.init();