+++ lib/dart_cqrs.dart
@@ -1,8 +1,7 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
+/// CQRS and event-driven architecture using get_it and injectable.
 library;
 
-export 'src/dart_cqrs_base.dart';
-
-// TODO: Export any libraries intended for clients of this package.
+export 'src/core/cqrs.dart';
+export 'src/core/dispatcher.dart';
+export 'src/injection.dart';
+export 'src/injection.module.dart';